@@ -1,6 +1,6 @@
 package com.chii.tt2info.ui;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.chii.tt2info.R;
